# openapi.yaml
openapi: 3.0.0
info:
  title: Budget Management API
  description: A comprehensive backend API for managing budgets, expenses, orders, transactions, and notifications.
  version: 1.0.0
  contact:
    name: Son Nguyen
    url: https://sonnguyenhoang.com
    email: hoangson091104@gmail.com
servers:
  - url: https://mongo-redis-rabbitmq-kafka-elk-backend.onrender.com
    description: Production Server
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - username
        - email
        - password
      properties:
        _id:
          type: string
          description: The unique identifier for the user.
        username:
          type: string
          description: The username of the user.
        email:
          type: string
          description: The email address of the user.
        password:
          type: string
          description: The hashed password of the user.
        createdAt:
          type: string
          format: date-time
          description: When the user was created.
    Budget:
      type: object
      required:
        - name
        - limit
      properties:
        _id:
          type: string
          description: The unique identifier for the budget.
        name:
          type: string
          description: The name of the budget.
        limit:
          type: number
          description: The spending limit for the budget.
        createdAt:
          type: string
          format: date-time
          description: When the budget was created.
...
    Expense:
      type: object
      required:
        - budgetId
        - description
        - amount
      properties:
        _id:
          type: string
          description: The unique identifier for the expense.
        budgetId:
          type: string
          description: The ID of the associated budget.
        description:
          type: string
          description: Details about the expense.
        amount:
          type: number
          description: The amount of the expense.
        createdAt:
          type: string
          format: date-time
          description: When the expense was created.
...
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: User registered successfully.
  /api/auth/login:
    post:
      summary: Authenticate a user and return a JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: User authenticated successfully.
...
  /api/budgets:
    post:
      summary: Create a new budget
      tags: [Budgets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Budget'
      responses:
        201:
          description: Budget created successfully.
    get:
      summary: Get all budgets
      tags: [Budgets]
      responses:
        200:
          description: List of budgets.
  /api/budgets/{id}:
    get:
      summary: Get a budget by ID
      tags: [Budgets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Budget retrieved successfully.
...
  /api/expenses:
    post:
      summary: Add a new expense to a budget
      tags: [Expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Expense'
      responses:
        201:
          description: Expense created successfully.
  /api/expenses/{budgetId}:
    get:
      summary: Retrieve all expenses for a specific budget
      tags: [Expenses]
      parameters:
        - name: budgetId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of expenses.
  /api/notifications/send:
    post:
      summary: Send a real-time notification
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        200:
          description: Notification sent successfully.
...